refactor struct
